meeting-reports/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
pydub==0.25.1
python-dotenv==1.0.0
langchain==0.0.350